refactor: move attach tests to own file
