+++ packages/rich-text-editor/test/attach-lit.test.js
@@ -1,0 +1,3 @@
+import '../theme/lumo/vaadin-rich-text-editor-styles.js';
+import '../src/vaadin-lit-rich-text-editor.js';
+import './attach.common.js';
