+++ packages/rich-text-editor/test/attach-polymer.test.js
@@ -1,0 +1,3 @@
+import '../theme/lumo/vaadin-rich-text-editor-styles.js';
+import '../src/vaadin-rich-text-editor.js';
+import './attach.common.js';
